Trocar tipo das variaveis de Float para Double

+++ src/main/java/tppe/tp1/acesso/Acesso.java
@@ -10,8 +10,8 @@ public class Acesso {
 	private LocalDateTime horaEntrada;
 	private LocalDateTime horaSaida;
 	private String tipoAcesso;
-	private Float valorAcesso;
-	private Float valorContratante;
+	private Double valorAcesso;
+	private Double valorContratante;
 
 	Acesso(AcessoBuilder builder){
 		placa = builder.getPlaca();
@@ -39,11 +39,11 @@ public class Acesso {
 		return tipoAcesso;
 	}
 
-	public Float getValorAcesso() {
+	public Double getValorAcesso() {
 		return valorAcesso;
 	}
 
-	public Float getValorContratante() {
+	public Double getValorContratante() {
 		return valorContratante;
 	}
